remove dead code from create
