+++ src/components/formItems/ProjectDiscountRate.tsx
@@ -1,8 +1,8 @@
 import { Trans } from '@lingui/macro'
 import { Form, Switch } from 'antd'
 
+import FormItemLabel from 'components/FormItemLabel'
 import { ThemeContext } from 'contexts/themeContext'
-import FormItemLabel from 'pages/create/FormItemLabel'
 import { CSSProperties, useContext } from 'react'
 import { defaultFundingCycleData } from 'redux/slices/editingV2Project'
