+++ src/components/formItems/ProjectReserved.tsx
@@ -1,8 +1,8 @@
 import { t, Trans } from '@lingui/macro'
 import { Form } from 'antd'
+import FormItemLabel from 'components/FormItemLabel'
 import { ThemeContext } from 'contexts/themeContext'
 import round from 'lodash/round'
-import FormItemLabel from 'pages/create/FormItemLabel'
 import { CSSProperties, useContext, useState } from 'react'
 import { defaultFundingCycleMetadata } from 'redux/slices/editingV2Project'
 import { formattedNum } from 'utils/format/formatNumber'
