+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/DistributionSplitModal/index.ts
@@ -1,0 +1,1 @@
+export * from './DistributionSplitModal'
