+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/FundingForm/DistributionSplitsSection/PayoutConfigurationExplainerCollapse.tsx
