+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/FundingForm/DistributionSplitsSection/index.ts
@@ -1,0 +1,1 @@
+export * from './DistributionSplitsSection'
