+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/RulesDrawer/RulesForm/RulesForm.tsx
@@ -15,13 +15,13 @@ import { BallotStrategy } from 'models/ballot'
 
 import isEqual from 'lodash/isEqual'
 
+import FormItemLabel from 'components/FormItemLabel'
 import { shadowCard } from 'constants/styles/shadowCard'
 import {
   ballotStrategies,
   DEFAULT_BALLOT_STRATEGY,
 } from 'constants/v2v3/ballotStrategies'
 import { getBallotStrategyByAddress } from 'utils/v2v3/ballotStrategies'
-import FormItemLabel from '../../FormItemLabel'
 import TokenMintingExtra from './TokenMintingExtra'
 
 export default function RulesForm({
