--- src/pages/create/models.ts
@@ -1,3 +1,0 @@
-export type TabContentProps = {
-  onFinish: VoidFunction
-}
