--- src/pages/create/tabs/ReviewDeployTab/utils.ts
@@ -1,17 +1,0 @@
-import { readProvider } from 'constants/readProvider'
-
-export const findTransactionReceipt = async (txHash: string) => {
-  let retries = 5
-  let receipt
-  while (retries > 0 && !receipt) {
-    receipt = await readProvider.getTransactionReceipt(txHash)
-    if (receipt) break
-
-    retries -= 1
-    // wait 2s
-    await new Promise(r => setTimeout(r, 2000))
-    console.info('Retrying tx receipt lookup...')
-  }
-
-  return receipt
-}
