+++ src/utils/nftFundingCycleMetadataOverrides.ts
@@ -1,0 +1,4 @@
+// Projects with NFT rewards need useDataSourceForPay to be true for NFT rewards to work
+export const NFT_FUNDING_CYCLE_METADATA_OVERRIDES = {
+  useDataSourceForPay: true,
+}
