--- src/utils/orEmpty.ts
@@ -1,1 +1,0 @@
-export const orEmpty = (str?: string) => (str && str.trim().length ? str : '--')
